# api/openapi.yaml
openapi: 3.1.0
info:
  title: API kiểm tra khối lượng hiện trường
  version: 0.6.0
  description: Hợp đồng API đã triển khai đến hết Mốc 5, gồm đối chiếu, snapshot và Excel/GeoJSON.
  license:
    name: Proprietary — sử dụng nội bộ/cá nhân
    identifier: LicenseRef-Proprietary
servers:
  - url: /api/v1
security:
  - cookieAuth: []
paths:
...
  /cases:
    get:
      tags: [cases]
      summary: Liệt kê hồ sơ thuộc người dùng hiện tại
      parameters:
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 100, default: 50 }
        - name: search
          in: query
          schema: { type: string, maxLength: 200 }
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/CaseStatus' }
      responses:
        '200':
          description: Danh sách hồ sơ
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/InspectionCase' }
                  nextCursor: { type: [string, 'null'] }
        default: { $ref: '#/components/responses/Error' }
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [cases]
      summary: Tạo hồ sơ và snapshot ranh giới địa bàn
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateCase' }
      responses:
        '201':
          description: Hồ sơ vừa tạo
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionCase' }
        default: { $ref: '#/components/responses/Error' }
  /cases/{caseId}:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [cases]
      summary: Lấy hồ sơ
      responses:
        '200':
          description: Hồ sơ
          headers:
            ETag: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionCase' }
        default: { $ref: '#/components/responses/Error' }
    patch:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [cases]
      summary: Sửa hồ sơ bằng optimistic concurrency
      parameters:
        - $ref: '#/components/parameters/IfMatch'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name: { type: string }
                periodStart: { type: string, format: date }
                periodEnd: { type: string, format: date }
                inspectedEntity: { type: [string, 'null'] }
                description: { type: [string, 'null'] }
                status: { enum: [draft, in_progress, review, archived] }
      responses:
        '200':
          description: Hồ sơ đã sửa
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InspectionCase' }
        '409': { $ref: '#/components/responses/Error' }
        default: { $ref: '#/components/responses/Error' }
    delete:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [cases]
      summary: Xóa mềm hồ sơ
      responses:
        '204': { description: Đã xóa mềm }
        default: { $ref: '#/components/responses/Error' }
  /cases/{caseId}/work-items:
    parameters:
      - $ref: '#/components/parameters/CaseId'
    get:
      tags: [cases]
      summary: Liệt kê công tác trong hồ sơ
      responses:
        '200':
          description: Danh sách công tác
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/WorkItem' }
        default: { $ref: '#/components/responses/Error' }
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [cases]
      summary: Thêm công tác và snapshot công thức
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [workTypeId, name]
              additionalProperties: false
              properties:
                workTypeId: { type: string, format: uuid }
                name: { type: string, minLength: 1, maxLength: 300 }
                periodStart: { type: [string, 'null'], format: date }
                periodEnd: { type: [string, 'null'], format: date }
                warningThreshold: { type: object }
      responses:
        '201':
          description: Công tác vừa thêm
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WorkItem' }
        default: { $ref: '#/components/responses/Error' }
  /cases/{caseId}/map-context:
    get:
      tags: [cases, measurements]
      summary: Lấy ranh giới snapshot làm ngữ cảnh bản đồ của hồ sơ
      parameters:
        - $ref: '#/components/parameters/CaseId'
      responses:
        '200':
          description: Ranh giới hồ sơ ở EPSG:4326
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CaseMapContext' }
        default: { $ref: '#/components/responses/Error' }
...
  /work-items/{workItemId}/routes:
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [routing]
      summary: Backend tính lại và lưu route chính thức
      responses:
        '201': { description: Route và measurement đã xác nhận }
        default: { $ref: '#/components/responses/Error' }
...
  /work-items/{workItemId}/gps-tracks:
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [field]
      summary: Lưu GPS raw/normalized bằng mutation idempotent
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
        - { name: X-Device-Id, in: header, required: true, schema: { type: string } }
      responses:
        '201': { description: GPS track mới }
        '200': { description: Kết quả cũ của cùng mutation }
        default: { $ref: '#/components/responses/Error' }
  /work-items/{workItemId}/gps-points:
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [field]
      summary: Ghi vị trí GPS hiện tại kèm accuracy bằng mutation idempotent
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string } }
        - { name: X-Device-Id, in: header, required: true, schema: { type: string } }
      responses:
        '201': { description: GPS point mới }
        '200': { description: Kết quả cũ của cùng mutation }
        default: { $ref: '#/components/responses/Error' }
  /work-items/{workItemId}/attachments:
    get:
      tags: [field]
      summary: Liệt kê ảnh đã xác minh của công tác để liên kết nguồn hoặc giải trình
      responses:
        '200': { description: Danh sách attachment completed thuộc công tác }
        default: { $ref: '#/components/responses/Error' }
  /attachments/presign:
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [field]
      summary: Tạo upload ảnh pending và URL MinIO có thời hạn
      responses:
        '201': { description: Upload pending }
        default: { $ref: '#/components/responses/Error' }
  /attachments/complete:
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [field]
      summary: Xác minh object và hoàn tất bằng chứng ảnh
      responses:
        '200': { description: Ảnh đã xác minh SHA-256 }
        default: { $ref: '#/components/responses/Error' }
  /work-items/{workItemId}/source-quantities:
    post:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [comparison]
      summary: Thêm khối lượng nguồn cho công tác chưa khóa
      responses:
        '201': { description: Khối lượng nguồn đã lưu }
        default: { $ref: '#/components/responses/Error' }
  /cases/{caseId}/comparison:
    get:
      tags: [comparison]
      summary: Tổng hợp measurement confirmed và đối chiếu từng nguồn
      responses:
        '200': { description: Chi tiết và tổng hợp theo nhóm/hồ sơ }
        default: { $ref: '#/components/responses/Error' }
  /cases/{caseId}/comparison-settings:
    patch:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [comparison]
      summary: Cấu hình ngưỡng tuyệt đối và tỷ lệ cấp hồ sơ
      responses:
        '200': { description: Ngưỡng đã lưu }
        default: { $ref: '#/components/responses/Error' }
  /source-quantities/{id}/explanation:
    put:
      security: [{ cookieAuth: [], csrfToken: [] }]
      tags: [comparison]
      summary: Lưu giải trình chênh lệch và attachment liên kết nếu có
      responses:
        '200': { description: Giải trình đã lưu }
        default: { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: dove_session
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
  parameters:
    CaseId:
      name: caseId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    IfMatch:
      name: If-Match
      in: header
      required: true
      schema: { type: string }
...
  responses:
    Error:
      description: Lỗi chuẩn có mã và trace ID
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [code, message, traceId]
      properties:
        code: { type: string }
        message: { type: string }
        details: { type: object }
        traceId: { type: string }
...
    CaseStatus:
      enum: [draft, in_progress, review, locked, archived]
    InspectionCase:
      type: object
      required: [id, caseCode, name, adminAreaId, adminAreaName, periodStart, periodEnd, status, workItemCount, version, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        caseCode: { type: string }
        name: { type: string }
        adminAreaId: { type: string, format: uuid }
        adminAreaName: { type: string }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        inspectedEntity: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/CaseStatus' }
        workItemCount: { type: integer, minimum: 0 }
        version: { type: integer, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateCase:
      type: object
      required: [caseCode, name, adminAreaId, periodStart, periodEnd]
      additionalProperties: false
      properties:
        caseCode: { type: string }
        name: { type: string }
        adminAreaId: { type: string, format: uuid }
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        inspectedEntity: { type: [string, 'null'] }
        description: { type: [string, 'null'] }
    WorkItem:
      type: object
      required: [id, caseId, workTypeId, workTypeCode, name, periodStart, periodEnd, unit, formulaSnapshot, warningThreshold, status]
      properties:
        id: { type: string, format: uuid }
        caseId: { type: string, format: uuid }
        workTypeId: { type: string, format: uuid }
        workTypeCode: { type: string }
        name: { type: string }
        periodStart: { type: [string, 'null'], format: date }
        periodEnd: { type: [string, 'null'], format: date }
        unit: { type: string }
        formulaSnapshot: { type: object }
        warningThreshold: { type: object }
        status: { enum: [draft, active, completed, archived] }
    GeoJsonGeometry:
      type: object
      required: [type, coordinates]
      additionalProperties: false
      properties:
        type: { enum: [Point, MultiPoint, LineString, MultiLineString, Polygon, MultiPolygon] }
        coordinates: {}
...
    CaseMapContext:
      type: object
      required: [caseId, boundary]
      properties:
        caseId: { type: string, format: uuid }
        boundary: { $ref: '#/components/schemas/GeoJsonGeometry' }
...